Use Exponential Scale for Shard Size Heat Map
According to Fechner's law subjectively perceived brightness is
proportional to the logarithm of the physical luminance of a color. Knowing
that it’s better to use an exponential scale when producing the colors for
the heat map so that differences are brought out more.

Props to @gibrown for the idea.